refactor(wasm): clean and remove unnecessary code
