+++ wasm/consensus/src/tasks/bridge.rs
@@ -1,7 +1,17 @@
 use clap::Args;
 use seda_runtime_sdk::{
     log,
-    wasm::{call_self, chain_call, chain_view, get_oracle_contract_id, memory_read, memory_write, Promise},
+    wasm::{
+        call_self,
+        chain_call,
+        chain_view,
+        get_oracle_contract_id,
+        memory_read,
+        memory_write,
+        shared_memory_get,
+        shared_memory_set,
+        Promise,
+    },
     Chain,
     FromBytes,
     Level,
@@ -21,6 +31,11 @@ pub struct Bridge {
 impl Bridge {
     pub fn handle(self) {
         log!(Level::Debug, "Bridge Handle");
+
+        // TODO: Temp fix, need to fix env variables
+        let contract_id = get_oracle_contract_id();
+        shared_memory_set("contract_id", contract_id.into());
+
         // we have a method to auto convert to bytes in a trait in runtime.
         // it should be moved to the sdk
         // TODO: SEDA-188 will make it so we can pass these instead of a vec of strings
@@ -35,6 +50,8 @@ impl Bridge {
 #[no_mangle]
 fn bridge_step_1() {
     log!(Level::Debug, "Bridge Step 1");
+    let contract_id =
+        String::from_utf8(shared_memory_get("contract_id")).expect("Could not read contract id from shared memory");
     let result = Promise::result(0);
     let deposit_bytes = memory_read("bridge_deposit");
     let deposit = u128::from_bytes_vec(deposit_bytes).unwrap();
@@ -46,7 +63,7 @@ fn bridge_step_1() {
             log!(Level::Debug, "Posting args: {args_string}");
             chain_call(
                 Chain::Near,
-                get_oracle_contract_id().as_str(), // TODO: Currently panics
+                contract_id, // TODO: Currently panics
                 "post_data_request",
                 args_string.into_bytes(),
                 deposit,
